api/pyproject.toml
[project]
name = "caprino-api"
version = "0.1.0"
description = "FastAPI project modeled from livestock management ERD (caprine/ovine)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.2.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1"
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
reload = true
